requirements.txt
python-dotenv
discord.py
aiohttp
colorlog
